+++ packages/@office-iss/react-native-win32/src-win/src/private/specs_DEPRECATED/modules/NativeImageLoaderWin32.js
@@ -14,7 +14,7 @@ import type {TurboModule} from '../../../../Libraries/TurboModule/RCTExport';
 import * as TurboModuleRegistry from '../../../../Libraries/TurboModule/TurboModuleRegistry';
 
 export interface Spec extends TurboModule {
-  +getConstants: () => {||};
+  +getConstants: () => {};
   // [Win32 uses callback instead of promise
   +getSize: (
     uri: string,