+++ packages/e2e-test-app-fabric/windows/RNTesterApp-Fabric/RNTesterApp-Fabric.cpp
@@ -365,6 +365,10 @@ void DumpUIAPatternInfo(IUIAutomationElement *pTarget, const winrt::Windows::Dat
   BOOL selectionRequired;
   BSTR text = nullptr;
   BOOL horizontallyScrollable;
+  BSTR annotationAuthor = nullptr;
+  BSTR annotationTypeName = nullptr;
+  BSTR annotationDateTime = nullptr;
+  int annotationTypeID = 0;
 
   // Dump IValueProvider Information
   IValueProvider *valuePattern;
@@ -467,6 +471,28 @@ void DumpUIAPatternInfo(IUIAutomationElement *pTarget, const winrt::Windows::Dat
     }
   }
 
+  // Dump IAnnotationProvider Information
+  winrt::com_ptr<IAnnotationProvider> annotationProvider;
+  hr = pTarget->GetCurrentPattern(UIA_AnnotationPatternId, reinterpret_cast<IUnknown **>(annotationProvider.put()));
+  if (SUCCEEDED(hr) && annotationProvider) {
+    hr = annotationProvider->get_AnnotationTypeId(&annotationTypeID);
+    if (SUCCEEDED(hr)) {
+      InsertIntValueIfNotDefault(result, L"AnnotationPattern.TypeId", annotationTypeID, 0);
+    }
+    hr = annotationProvider->get_AnnotationTypeName(&annotationTypeName);
+    if (SUCCEEDED(hr)) {
+      InsertStringValueIfNotEmpty(result, L"AnnotationPattern.TypeName", annotationTypeName);
+    }
+    hr = annotationProvider->get_Author(&annotationAuthor);
+    if (SUCCEEDED(hr)) {
+      InsertStringValueIfNotEmpty(result, L"AnnotationPattern.Author", annotationAuthor);
+    }
+    hr = annotationProvider->get_DateTime(&annotationDateTime);
+    if (SUCCEEDED(hr)) {
+      InsertStringValueIfNotEmpty(result, L"AnnotationPattern.DateTime", annotationDateTime);
+    }
+  }
+
   // Dump IScrollProvider Information
   winrt::com_ptr<IScrollProvider> scrollPattern;
   hr = pTarget->GetCurrentPattern(UIA_ScrollPatternId, reinterpret_cast<IUnknown **>(scrollPattern.put()));
@@ -497,6 +523,8 @@ winrt::Windows::Data::Json::JsonObject DumpUIATreeRecurse(
   int level = 0;
   LiveSetting liveSetting = LiveSetting::Off;
   BSTR itemStatus;
+  BSTR itemType;
+  BSTR accessKey;
 
   pTarget->get_CurrentAutomationId(&automationId);
   pTarget->get_CurrentControlType(&controlType);
@@ -506,6 +534,8 @@ winrt::Windows::Data::Json::JsonObject DumpUIATreeRecurse(
   pTarget->get_CurrentLocalizedControlType(&localizedControlType);
   pTarget->get_CurrentName(&name);
   pTarget->get_CurrentItemStatus(&itemStatus);
+  pTarget->get_CurrentItemType(&itemType);
+  pTarget->get_CurrentAccessKey(&accessKey);
   IUIAutomationElement4 *pTarget4;
   HRESULT hr = pTarget->QueryInterface(__uuidof(IUIAutomationElement4), reinterpret_cast<void **>(&pTarget4));
   if (SUCCEEDED(hr) && pTarget4) {
@@ -528,6 +558,8 @@ winrt::Windows::Data::Json::JsonObject DumpUIATreeRecurse(
   InsertIntValueIfNotDefault(result, L"Level", level);
   InsertLiveSettingValueIfNotDefault(result, L"LiveSetting", liveSetting);
   InsertStringValueIfNotEmpty(result, L"ItemStatus", itemStatus);
+  InsertStringValueIfNotEmpty(result, L"ItemType", itemType);
+  InsertStringValueIfNotEmpty(result, L"AccessKey", accessKey);
   DumpUIAPatternInfo(pTarget, result);
 
   IUIAutomationElement *pChild;