+++ vnext/Microsoft.ReactNative/ReactHost/ReactHost.cpp
@@ -6,6 +6,8 @@
 #include <ReactPropertyBag.h>
 #include <winrt/Windows.Foundation.h>
 
+#include <CppRuntimeOptions.h>
+
 #include <react/featureflags/ReactNativeFeatureFlags.h>
 #include <react/featureflags/ReactNativeFeatureFlagsDefaults.h>
 
@@ -282,8 +284,12 @@ bool ReactOptions::EnableDefaultCrashHandler() const noexcept {
 
 class ReactNativeWindowsFeatureFlags : public facebook::react::ReactNativeFeatureFlagsDefaults {
  public:
-  bool disableEventLoopOnBridgeless() override {
-    return true; // Disable event loop until we are on a JSI version that supports microtasks
+  bool enableBridgelessArchitecture() override {
+#ifdef USE_FABRIC
+    return true;
+#else
+    return false;
+#endif
   }
 
   bool enableCppPropsIteratorSetter() override {