+++ vnext/src-win/Libraries/Components/Pressable/Pressable.windows.js
@@ -39,11 +39,11 @@ import type {HandledKeyboardEvent} from '../../Components/View/ViewPropTypes';
 
 type ViewStyleProp = $ElementType<React.ElementConfig<typeof View>, 'style'>;
 
-export type StateCallbackType = $ReadOnly<{|
+export type StateCallbackType = $ReadOnly<{
   pressed: boolean,
-|}>;
+}>;
 
-type Props = $ReadOnly<{|
+type Props = $ReadOnly<{
   /**
    * Accessibility.
    */
@@ -241,7 +241,7 @@ type Props = $ReadOnly<{|
    * https://github.com/facebook/react-native/issues/34424
    */
   'aria-label'?: ?string,
-|}>;
+}>;
 
 type Instance = React.ElementRef<typeof View>;