+++ packages/@react-native-windows/automation/CHANGELOG.md
@@ -1,8 +1,38 @@
 # Change Log - @react-native-windows/automation
 
-<!-- This log was last generated on Thu, 01 May 2025 05:24:00 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 13 May 2025 21:05:07 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.3.375
+
+Tue, 13 May 2025 21:05:07 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.293
+- Bump @react-native-windows/fs to v0.0.0-canary.64
+- Bump @rnw-scripts/eslint-config to v1.2.35
+- Bump @rnw-scripts/just-task to v2.3.52
+
+## 0.3.374
+
+Thu, 08 May 2025 05:23:33 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.292
+- Bump @react-native-windows/fs to v0.0.0-canary.63
+- Bump @rnw-scripts/eslint-config to v1.2.34
+- Bump @rnw-scripts/just-task to v2.3.51
+
+## 0.3.373
+
+Wed, 07 May 2025 05:27:07 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.291
 
 ## 0.3.372